feat(fs): add /proc/mounts, /proc/kmsg, disk partition devnodes, and unmount improvements

+++ fs/fat32.h
@@ -106,7 +106,7 @@ typedef fat32_file_handle_t FAT32_FileHandle;
 
 typedef struct {
     char name[FAT32_MAX_FILENAME];
-    uint32_t size;
+    uint64_t size;
     uint8_t is_directory;
     uint32_t start_cluster;
     uint16_t write_date;
